FeatureHouse for FOSD
Using FeatureIDE that is an eclipse-based IDE, implemented a graph library in order to generate the variants of a program or product. A variant from the software product line can be generated using the configuration tool of Feature IDE. Features are implemented and reusable components developed to meet the Feature oriented software development (FOSD) standards, that helps in the incremental development or enhancement of the programs. It also makes a clear way for the program generation in the software product line (SPL).

FeatureIDE, FeatureHouse
Graph library, Graph product line
Java 8
Eclipse